Add 3d fdtd. Remove macros for inlines. Need testing on ABC

+++ fdtdCore/fdtd-grid.h
@@ -8,17 +8,17 @@ enum GRIDTYPE {oneDGrid, teZGrid, tmZGrid, threeDGrid};
 
 //FDTD grid containing all field cells and time data
 struct Grid {
-	double		*ex, *cexe, *cexh;		//Ex field
-	double		*ey, *ceye, *ceyh;		//Ey field
-	double		*ez, *ceze, *cezh;		//Ez field
-	double		*hx, *chxh, *chxe;		//Hx field
-	double		*hy, *chyh, *chye;		//Hy field
-	double		*hz, *chzh, *chze;		//Hz field
+	float		*ex, *cexe, *cexh;		//Ex field
+	float		*ey, *ceye, *ceyh;		//Ey field
+	float		*ez, *ceze, *cezh;		//Ez field
+	float		*hx, *chxh, *chxe;		//Hx field
+	float		*hy, *chyh, *chye;		//Hy field
+	float		*hz, *chzh, *chze;		//Hz field
 	int			sizeX, sizeY, sizeZ;	//grid size
 	int			time;		//current timestep
 	int			maxTime;	//final timestep
 	int			type;		// use with GRIDTYPE for 1D, 2D or 3D grids
-	double		cdtds;		//courant number
+	float		cdtds;		//courant number
 };
 typedef struct Grid Grid;
